modernize test suite of simiulating jumps
